requirements.txt
asgiref==3.9.1
attrs==25.3.0
click==8.2.1
colorama==0.4.6
croniter==6.0.0
Django==5.2.5
django-filter==25.1
django-redis==6.0.0
django-rq==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docker-composer==2.79.7
loguru==0.7.3
pillow==11.3.0
prometheus_client==0.23.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==6.4.0
redis-cli==1.0.1
rq==2.6.0
ruff==0.12.11
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
win32_setctime==1.2.0